+++ examples/repeat.rs
@@ -1,4 +1,4 @@
-use figura::{Context, DefaultParser, Template, Value};
+use figura::{Context, Template, Value};
 
 fn main() {
     let mut ctx = Context::new();
@@ -7,7 +7,7 @@ fn main() {
     ctx.insert("title", Value::static_str("HEADER"));
 
     let template =
-        Template::<'{', '}'>::compile::<DefaultParser>("{char:count}\n{title}\n{char:count}")
+        Template::<'{', '}'>::compile("{char:count}\n{title}\n{char:count}")
             .unwrap();
 
     println!("{}", template.format(&ctx).unwrap());
@@ -16,12 +16,12 @@ fn main() {
     ctx.insert("times", Value::Int(10));
 
     let template =
-        Template::<'{', '}'>::compile::<DefaultParser>("Pattern: {pattern:times}").unwrap();
+        Template::<'{', '}'>::compile("Pattern: {pattern:times}").unwrap();
 
     println!("{}", template.format(&ctx).unwrap());
 
     let template =
-        Template::<'{', '}'>::compile::<DefaultParser>("{'#':5} Progress Bar {'#':5}").unwrap();
+        Template::<'{', '}'>::compile("{'#':5} Progress Bar {'#':5}").unwrap();
 
     println!("{}", template.format(&Context::new()).unwrap());
 }